Replace c-style arrays with std::array where possible

+++ src/fileio.h
@@ -142,7 +142,7 @@ typedef struct VIRTUALFILE
     VIRTUALTYPE             m_type;                                 /**< @brief Type of this virtual file */
     int                     m_flags;                                /**< @brief One of the VIRTUALFLAG_* flags */
 
-    int                     m_format_idx;                           /**< @brief Index into params.format[] array */
+    size_t                  m_format_idx;                           /**< @brief Index into params.format[] array */
     std::string             m_destfile;                             /**< @brief Name and path of destination file */
     std::string             m_virtfile;                             /**< @brief Name and path of virtual file */
     std::string             m_origfile;                             /**< @brief Sanitised name and path of original file */
